+++ shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/area/AreaPresenter.kt
@@ -73,13 +73,13 @@ internal class AreaPresenter(
         var title by rememberSaveable { mutableStateOf(screen.title.orEmpty()) }
         var isError by rememberSaveable { mutableStateOf(false) }
         var recordedHistory by rememberSaveable { mutableStateOf(false) }
-        val topAppBarFilterState = rememberTopAppBarFilterState()
-        val query = topAppBarFilterState.filterText
         var area: AreaDetailsModel? by rememberRetained { mutableStateOf(null) }
         val tabs: List<AreaTab> by rememberSaveable {
             mutableStateOf(AreaTab.entries)
         }
         var selectedTab by rememberSaveable { mutableStateOf(AreaTab.DETAILS) }
+        val topAppBarFilterState = rememberTopAppBarFilterState()
+        val query = topAppBarFilterState.filterText
         var forceRefreshDetails by remember { mutableStateOf(false) }
         val detailsLazyListState = rememberLazyListState()
         var snackbarMessage: String? by rememberSaveable { mutableStateOf(null) }
@@ -143,6 +143,11 @@ internal class AreaPresenter(
             key1 = query,
             key2 = selectedTab,
         ) {
+            topAppBarFilterState.show(
+                selectedTab !in listOf(
+                    AreaTab.STATS,
+                )
+            )
             when (selectedTab) {
                 AreaTab.DETAILS -> {
                     // Loaded above
