+++ shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/area/AreaUi.kt
@@ -145,9 +145,6 @@ internal fun AreaUiInternal(
                 title = state.title,
                 scrollBehavior = scrollBehavior,
                 overflowDropdownMenuItems = overflowDropdownMenuItems,
-                showFilterIcon = state.selectedTab !in listOf(
-                    AreaTab.STATS,
-                ),
                 topAppBarFilterState = state.topAppBarFilterState,
                 additionalBar = {
                     TabsBar(
