+++ shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/instrument/InstrumentUi.kt
@@ -85,9 +85,6 @@ internal fun InstrumentUi(
                 entity = entity,
                 title = state.title,
                 scrollBehavior = scrollBehavior,
-                showFilterIcon = state.selectedTab !in listOf(
-                    InstrumentTab.STATS,
-                ),
                 overflowDropdownMenuItems = {
                     OpenInBrowserMenuItem(
                         url = state.url,
