+++ shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/recording/RecordingPresenter.kt
@@ -59,13 +59,13 @@ internal class RecordingPresenter(
         var subtitle by rememberSaveable { mutableStateOf("") }
         var isError by rememberSaveable { mutableStateOf(false) }
         var recordedHistory by rememberSaveable { mutableStateOf(false) }
-        val topAppBarFilterState = rememberTopAppBarFilterState()
-        val query = topAppBarFilterState.filterText
         var recording: RecordingDetailsModel? by rememberRetained { mutableStateOf(null) }
         val tabs: List<RecordingTab> by rememberSaveable {
             mutableStateOf(RecordingTab.entries)
         }
         var selectedTab by rememberSaveable { mutableStateOf(RecordingTab.DETAILS) }
+        val topAppBarFilterState = rememberTopAppBarFilterState()
+        val query = topAppBarFilterState.filterText
         var forceRefreshDetails by remember { mutableStateOf(false) }
         val detailsLazyListState = rememberLazyListState()
         var snackbarMessage: String? by rememberSaveable { mutableStateOf(null) }
@@ -121,6 +121,11 @@ internal class RecordingPresenter(
             key1 = query,
             key2 = selectedTab,
         ) {
+            topAppBarFilterState.show(
+                selectedTab !in listOf(
+                    RecordingTab.STATS,
+                )
+            )
             when (selectedTab) {
                 RecordingTab.DETAILS -> {
                     // Loaded above
