+++ shared/feature/images/src/androidUnitTest/kotlin/ly/david/musicsearch/shared/feature/images/CoverArtsUiTest.kt
@@ -3,7 +3,14 @@ package ly.david.musicsearch.shared.feature.images
 import ly.david.musicsearch.ui.test.screenshot.ScreenshotTest
 import org.junit.Test
 
-class CoverArtsPagerUiTest : ScreenshotTest() {
+class CoverArtsUiTest : ScreenshotTest() {
+
+    @Test
+    fun coverArtsGrid() {
+        snapshot {
+            PreviewCoverArtsGridUi()
+        }
+    }
 
     @Test
     fun coverArtsPagerCompact() {
