+++ ui/common/src/androidMain/kotlin/ly/david/musicsearch/ui/common/topappbar/TopAppBarWithFilter.android.kt
@@ -30,7 +30,6 @@ actual fun TopAppBarWithFilter(
     overflowDropdownMenuItems: @Composable (OverflowMenuScope.() -> Unit)?,
     subtitleDropdownMenuItems: @Composable (OverflowMenuScope.() -> Unit)?,
 
-    showFilterIcon: Boolean,
     topAppBarFilterState: TopAppBarFilterState,
     topAppBarEditState: TopAppBarEditState,
 
@@ -57,7 +56,6 @@ actual fun TopAppBarWithFilter(
         scrollBehavior = scrollBehavior,
         overflowDropdownMenuItems = overflowDropdownMenuItems,
         subtitleDropdownMenuItems = subtitleDropdownMenuItems,
-        showFilterIcon = showFilterIcon,
         topAppBarFilterState = topAppBarFilterState,
         topAppBarEditState = topAppBarEditState,
         additionalActions = additionalActions,
@@ -109,7 +107,9 @@ internal fun PreviewTopAppBarWithFilterNoFilter() {
     PreviewTheme {
         TopAppBarWithFilterInternal(
             title = "Title",
-            showFilterIcon = false,
+            topAppBarFilterState = TopAppBarFilterState(
+                initialShow = false,
+            )
         )
     }
 }
