+++ ui/common/src/iosMain/kotlin/ly/david/musicsearch/ui/common/topappbar/TopAppBarWithFilter.ios.kt
@@ -23,7 +23,6 @@ actual fun TopAppBarWithFilter(
     overflowDropdownMenuItems: @Composable (OverflowMenuScope.() -> Unit)?,
     subtitleDropdownMenuItems: @Composable (OverflowMenuScope.() -> Unit)?,
 
-    showFilterIcon: Boolean,
     topAppBarFilterState: TopAppBarFilterState,
     topAppBarEditState: TopAppBarEditState,
 
@@ -40,7 +39,6 @@ actual fun TopAppBarWithFilter(
         scrollBehavior = scrollBehavior,
         overflowDropdownMenuItems = overflowDropdownMenuItems,
         subtitleDropdownMenuItems = subtitleDropdownMenuItems,
-        showFilterIcon = showFilterIcon,
         topAppBarFilterState = topAppBarFilterState,
         topAppBarEditState = topAppBarEditState,
         additionalActions = additionalActions,
